Refactor Page Objects and Blog Demo
The Page Objects where using specific SeleniumBase commands, these makes the Page Objects not reusable, if one wanted to use them for E2E tests.

Hence, this commit includes a new Driver interface that can be injected into the Page Objects and tha provides the methods that Page Objects need. With this, a SeleniumBase demo and an E2E test can reuse the same Page Object without the need to know the details of what methods are used by each driver.