Use SQLite also for Dev Settings
Since these are the Settings that are used for Preview Apps, and the Preview Apps have a dynamic / temporary PostgreSQL instance, we can't use a static DATABASE_URL reference.
Having an SQLite for Preview Apps would work, since with this, we only want to verify that the app is successfully deployed and that it can start with no errors.